+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 4.1.1 under development
 
+- New #277: Add `beforeCheckbox()` and `afterCheckbox()` methods to `CheckboxList`, and `beforeRadio()` and
+  `afterRadio()` methods to `RadioList` (@vjik)
 - New #276: Add `beforeInput()` and `afterInput()` methods to abstract `BooleanInputTag`, extended by `Radio`
   and `Checkbox` (@vjik)